Shrink idea of "Reactors"
A Reactor can be waaay less complex if we simply use it as "map"-helper
for mapping over `Stream`s.

If we map over a stream of Vec<u8> and deserialize them to
GossipMessages in one step, and handle them appropriately in the next
step, it is way less complex to implement these things and we do not
have to care about this whole "how do I shut down the thing" because we
can simply drop() everything and let the destructors do their job.

This patch removes the Reactor nonsense.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>